requirements.txt
discord.py==2.0.0
yapf==0.32.0